+++ common/src/states/climb.rs
@@ -79,7 +79,7 @@ impl CharacterBehavior for Data {
                     CLIMB_BOOST_JUMP_FACTOR * impulse / data.mass.0,
                 ));
             };
-            update.character = CharacterState::Idle(idle::Data { is_sneaking: false });
+            update.character = CharacterState::Idle(idle::Data::default());
             return update;
         };
         // Move player
@@ -103,7 +103,7 @@ impl CharacterBehavior for Data {
             .try_change_by(-energy_use * data.dt.0)
             .is_err()
         {
-            update.character = CharacterState::Idle(idle::Data { is_sneaking: false });
+            update.character = CharacterState::Idle(idle::Data::default());
         }
 
         // Set orientation direction based on wall direction